+++ docs/docs_src/quickstart/documentation/custom_schema.py
@@ -1,7 +1,7 @@
 from propan import PropanApp, RabbitBroker
 from propan.asyncapi.main import AsyncAPISchema
 from propan.cli.docs.gen import gen_app_schema_json, gen_app_schema_yaml, get_app_schema
-from propan.cli.docs.serve import get_asyncapi_html
+from propan.cli.docs.serving import get_asyncapi_html
 
 broker = RabbitBroker()
 app = PropanApp(broker)
